Modularized the design and made the tree use es6 classes. It's much clearner now
