+++ hatchet/vis/webpack.config.js
@@ -28,7 +28,7 @@ module.exports = {
         path: path.resolve(__dirname, 'static')
     },
     optimization: {
-        minimize: true
+        minimize: false
     },
     plugins:[
         new HtmlWebpackPlugin({
